Make Clippy pass
Mine:
- Redundant u16 casts; the policy newtypes already hold u16.
- mem::forget on the hydration Connection, which has no Drop, so it was a
  no-op dressed up as intent. The OS connection lives until
  CfDisconnectSyncRoot regardless; the handle now sits in a OnceLock so a
  future clean shutdown has something to close.
- Unused re-exports allowed, since they are the modules' surface.

Pre-existing, part of why CI has been red since June:
- Unused Serialize and RecommendedWatcher imports.
- unwrap_err after is_ok, replaced with err().cloned().
- Needless borrows, consecutive str::replace, manual checked division.
- Deserialise-only structs marked allow(dead_code); the fields mirror the API
  payload whether or not anything reads them yet.

+++ src-tauri/src/cloudfiles.rs
@@ -97,9 +97,9 @@ mod imp {
         let policies = CF_SYNC_POLICIES {
             StructSize: size_of::<CF_SYNC_POLICIES>() as u32,
             Hydration: CF_HYDRATION_POLICY {
-                Primary: CF_HYDRATION_POLICY_PRIMARY(CF_HYDRATION_POLICY_FULL.0 as u16),
+                Primary: CF_HYDRATION_POLICY_PRIMARY(CF_HYDRATION_POLICY_FULL.0),
                 Modifier: CF_HYDRATION_POLICY_MODIFIER(
-                    CF_HYDRATION_POLICY_MODIFIER_AUTO_DEHYDRATION_ALLOWED.0 as u16,
+                    CF_HYDRATION_POLICY_MODIFIER_AUTO_DEHYDRATION_ALLOWED.0,
                 ),
             },
             // ALWAYS_FULL, not FULL. We enumerate the entire remote index and
@@ -113,7 +113,7 @@ mod imp {
             // open sat waiting for the timeout. ALWAYS_FULL tells the platform
             // never to forward enumeration at all.
             Population: CF_POPULATION_POLICY {
-                Primary: CF_POPULATION_POLICY_PRIMARY(CF_POPULATION_POLICY_ALWAYS_FULL.0 as u16),
+                Primary: CF_POPULATION_POLICY_PRIMARY(CF_POPULATION_POLICY_ALWAYS_FULL.0),
                 Modifier: CF_POPULATION_POLICY_MODIFIER(0),
             },
             InSync: CF_INSYNC_POLICY_TRACK_ALL,
@@ -234,4 +234,5 @@ mod imp {
     }
 }
 
+#[allow(unused_imports)]
 pub use imp::{create_placeholders, is_placeholder, is_supported, register, unregister};
